+++ src/apps/PreferencesWindow/PreferencesWindow/ComplexModificationsRulesTableViewDataSource.h
@@ -1,0 +1,6 @@
+// -*- Mode: objc -*-
+
+@import Cocoa;
+
+@interface ComplexModificationsRulesTableViewDataSource : NSObject <NSTableViewDataSource>
+@end