+++ src/apps/PreferencesWindow/PreferencesWindow/LogFileTextViewController.m
@@ -2,13 +2,21 @@
 #import "libkrbn.h"
 #import "weakify.h"
 
+typedef enum {
+  LogLevelInfo,
+  LogLevelWarn,
+  LogLevelErr,
+} LogLevel;
+
 @interface LogFileTextViewController ()
 
 @property(unsafe_unretained) IBOutlet NSTextView* textView;
 @property(weak) IBOutlet NSTextField* currentTime;
+@property(weak) IBOutlet NSTabViewItem* logTabViewItem;
 @property dispatch_source_t timer;
 @property NSDate* fileModificationDate;
 @property libkrbn_log_monitor* libkrbn_log_monitor;
+@property NSInteger errorLogCount;
 
 - (void)appendLogLine:(const char*)line;
 
@@ -38,11 +46,12 @@ static void log_updated_callback(const char* line, void* refcon) {
     size_t size = libkrbn_log_monitor_initial_lines_size(self.libkrbn_log_monitor);
     for (size_t i = 0; i < size; ++i) {
       const char* line = libkrbn_log_monitor_initial_line(self.libkrbn_log_monitor, i);
+      LogLevel level = [self getLogLevel:line];
       if (line) {
         [self.textView.textStorage appendAttributedString:[[NSAttributedString alloc] initWithString:[self logLineString:line]
-                                                                                          attributes:[self stringAttributes]]];
+                                                                                          attributes:[self stringAttributes:level]]];
         [self.textView.textStorage appendAttributedString:[[NSAttributedString alloc] initWithString:@"\n"
-                                                                                          attributes:[self stringAttributes]]];
+                                                                                          attributes:[self stringAttributes:level]]];
       }
     }
   }
@@ -73,6 +82,10 @@ static void log_updated_callback(const char* line, void* refcon) {
   }
 }
 
+- (void)resetErrorLogCount {
+  self.errorLogCount = 0;
+}
+
 - (void)dealloc {
   if (self.libkrbn_log_monitor) {
     libkrbn_log_monitor* p = self.libkrbn_log_monitor;
@@ -90,6 +103,11 @@ static void log_updated_callback(const char* line, void* refcon) {
 
 - (void)appendLogLine:(const char*)line {
   NSString* lineString = [self logLineString:line];
+  LogLevel level = [self getLogLevel:line];
+
+  if (level == LogLevelErr) {
+    ++(self.errorLogCount);
+  }
 
   @weakify(self);
   dispatch_async(dispatch_get_main_queue(), ^{
@@ -98,18 +116,80 @@ static void log_updated_callback(const char* line, void* refcon) {
 
     [self.textView.textStorage beginEditing];
     [self.textView.textStorage appendAttributedString:[[NSAttributedString alloc] initWithString:lineString
-                                                                                      attributes:[self stringAttributes]]];
+                                                                                      attributes:[self stringAttributes:level]]];
     [self.textView.textStorage appendAttributedString:[[NSAttributedString alloc] initWithString:@"\n"
-                                                                                      attributes:[self stringAttributes]]];
+                                                                                      attributes:[self stringAttributes:level]]];
     [self.textView.textStorage endEditing];
 
     [self scrollToBottom];
+    [self updateTabLabel];
   });
 }
 
-- (NSDictionary*)stringAttributes {
+- (void)updateTabLabel {
+  if (self.logTabViewItem.tabState == NSSelectedTab) {
+    [self resetErrorLogCount];
+  } else {
+    if (self.errorLogCount > 0) {
+      self.logTabViewItem.label = [NSString stringWithFormat:@"Log \u2666%ld", self.errorLogCount];
+      return;
+    }
+  }
+
+  self.logTabViewItem.label = @"Log";
+}
+
+- (LogLevel)getLogLevel:(const char*)line {
+  if (libkrbn_is_warn_log(line)) {
+    return LogLevelWarn;
+  }
+  if (libkrbn_is_err_log(line)) {
+    return LogLevelErr;
+  }
+  return LogLevelInfo;
+}
+
+- (NSColor*)getBackgroundColor:(LogLevel)level {
+  if (level == LogLevelWarn) {
+    // #fcf8e3
+    return [NSColor colorWithRed:(CGFloat)(0xfc) / 255
+                           green:(CGFloat)(0xf8) / 255
+                            blue:(CGFloat)(0xe3) / 255
+                           alpha:1.0];
+  }
+  if (level == LogLevelErr) {
+    // #f2dede
+    return [NSColor colorWithRed:(CGFloat)(0xf2) / 255
+                           green:(CGFloat)(0xde) / 255
+                            blue:(CGFloat)(0xde) / 255
+                           alpha:1.0];
+  }
+  return [NSColor textBackgroundColor];
+}
+
+- (NSColor*)getForegroundColor:(LogLevel)level {
+  if (level == LogLevelWarn) {
+    // #8a6d3b
+    return [NSColor colorWithRed:(CGFloat)(0x8a) / 255
+                           green:(CGFloat)(0x6d) / 255
+                            blue:(CGFloat)(0x3b) / 255
+                           alpha:1.0];
+  }
+  if (level == LogLevelErr) {
+    // #a94442
+    return [NSColor colorWithRed:(CGFloat)(0xa9) / 255
+                           green:(CGFloat)(0x44) / 255
+                            blue:(CGFloat)(0x42) / 255
+                           alpha:1.0];
+  }
+  return [NSColor textColor];
+}
+
+- (NSDictionary*)stringAttributes:(LogLevel)level {
   return @{
     NSFontAttributeName : [NSFont fontWithName:@"Menlo" size:11],
+    NSBackgroundColorAttributeName : [self getBackgroundColor:level],
+    NSForegroundColorAttributeName : [self getForegroundColor:level],
   };
 }