+++ src/apps/PreferencesWindow/PreferencesWindow/PreferencesWindowController.m
@@ -1,4 +1,5 @@
 #import "PreferencesWindowController.h"
+#import "ComplexModificationsRulesTableViewController.h"
 #import "DevicesTableViewController.h"
 #import "FnFunctionKeysTableViewController.h"
 #import "KarabinerKit/KarabinerKit.h"
@@ -13,6 +14,7 @@
 
 @interface PreferencesWindowController ()
 
+@property(weak) IBOutlet ComplexModificationsRulesTableViewController* complexModificationsRulesTableViewController;
 @property(weak) IBOutlet DevicesTableViewController* devicesTableViewController;
 @property(weak) IBOutlet FnFunctionKeysTableViewController* fnFunctionKeysTableViewController;
 @property(weak) IBOutlet LogFileTextViewController* logFileTextViewController;
@@ -45,6 +47,7 @@
   // Setup
 
   [self.fnFunctionKeysTableViewController setup];
+  [self.complexModificationsRulesTableViewController setup];
   [self.devicesTableViewController setup];
   [self setupVirtualHIDKeyboardTypePopUpButton];
   [self setupVirtualHIDKeyboardCapsLockDelayMilliseconds:nil];
@@ -114,6 +117,10 @@
 - (void)show {
   [self.window makeKeyAndOrderFront:self];
   [NSApp activateIgnoringOtherApps:YES];
+}
+
+- (void)tabView:(NSTabView*)tabView didSelectTabViewItem:(NSTabViewItem*)tabViewItem {
+  [self.logFileTextViewController updateTabLabel];
 }
 
 - (void)setupVirtualHIDKeyboardTypePopUpButton {